refactor: adjust API types, caching, and config

+++ src/api/WalletUp.Application/Abstractions/IExchangeRateService.cs
@@ -4,5 +4,5 @@ namespace WalletUp.Application.Abstractions;
 
 public interface IExchangeRateService
 {
-    Task<ExchangeApiResponse> GetRatesAsync(string currencies, decimal amount);
+    Task<ExchangeApiResponse> GetRatesAsync(string currencies, double amount);
 }
