+++ src/api/WalletUp.Infstructre/Refit/IExchangeApi.cs
@@ -7,6 +7,6 @@ namespace CashCat.Infstructre.Refit;
 public interface IExchangeApi
 {
     [Get("/api/currencyconverter/exchange")]
-    Task<ExchangeApiResponse> Convert([Query]  string currencies, [Query] decimal  amount);
+    Task<ExchangeApiResponse> Convert([Query]  string currencies, [Query] double  amount);
 
 }
